Fix function return + add to getHumanChoice
Fix return in function playRound, it was not logging the desired strings
in the console because the returned values were not being loged at any
point of the code, now instead of returning the result, it logs it into
the console.

Now the getHumanChoice function capitalizes the first letter of the
returned string/choice.